Updated misc. small details.
Changed the header. Went from using the Bootstrap grid to Bootstrap flex. Done to get the spacing to match the mockups, and to avoid all the extra html tags.
Changed the padding for the main content area to make it match the mockup more closely.
Changed cards to be of equal sizing when viewed beside each other.
Changed the layout inside the cards from Bootstrap flex to Bootstrap grid.